TreeLoaders

Documentation for TreeLoaders.

This package provides functions to load various datasets containing tree data into consistent DataFrames.DataFrames.

It as well provides functionality to calculate shadows from these datasets.

Interface

To properly work within the MinistryOfCoolWalks ecosystem, we expect the dataframes to follow a certain set of requirements:

Each row in the DataFrame represents a single tree

Each DataFrame must have the following columns:

:id unique id of the tree

:lon longitude of tree (WSG84)

:lat latitude of tree (WSG84)

:height height of tree from floor to center of crown (m)

:radius radius of crown (m)

:geometry ArchGDAL point, representing center of stem (must have spatial reference applied)

The DataFrame must have metadata with a key of observatory, which contains a CoolWalksUtils.ShadowObservatory. This value contains the center coordinates used for projection to a local coordinate system and the timezone of the dataset, used to calculate the sunposition, for shadow projection.

The technical part of the last two of these requirements can be checked with TreeLoaders.check_tree_dataframe_integrity(df).

API

TreeLoaders.check_tree_dataframe_integrity

TreeLoaders.check_tree_dataframe_integrity — Method

check_tree_dataframe_integrity(df)

Checks if df conforms to the technical requirements needed to be considered as a source for tree data.