Loading Nottingham Trees

Introduction

We load the tree data from the open data server of the city concil of nottingham, available here: nottingham tree data.

As there are a lot of duplicates in this dataset, we do some preliminary cleanup in load_nottingham_trees. During this process, we also drop a lot of irrelevant columns and create a few new ones. All trees which are (transitively) within a square of 2e-5 meter edge length around a tree are considered the same tree. We average over all numerical data and keep the first entry for every other one.

API

TreeLoaders.load_nottingham_trees

TreeLoaders.load_nottingham_trees — Method

load_nottingham_trees(path; extent=nothing)

Loads tree data from nottingham. Downloaded from https://maps164.nottinghamcity.gov.uk/server/rest/services/OpenData/OpenData/MapServer/91.

merges the considerable amount of duplicate trees in the original dataset by distance (trees closer than 1e-5m are considered equal)

arguments

path: Path to the file with the dataset

extent: Extents.Extent, specifying a clipping range for the Dataset. Use X for lon and Y for lat.

returns

DataFrame conforming to the requirements needed to be considered a source of tree data.