Loading Spanish Trees

Barcelona

We load the tree data from the open data server of barcelona. This Dataset comes in three different subfiles:

Street Trees

Park Trees

Zone Trees

Since there are no height and radius data available, we set them constant to height=6m and radius=3.5m. These values are the mean values found in the Nottingham dataset.

Valencia

We load the tree data from the open data server of valencia, available here: valencia tree data.

Since there are no height and radius data available, we set them constant to height=6m and radius=3.5m. These values are the mean values found in the Nottingham dataset.

API

TreeLoaders.load_barcelona_trees

TreeLoaders.load_valencia_trees

TreeLoaders.load_barcelona_trees — Method

load_barcelona_trees(path; extent=nothing)

Loads tree data from barcelona. Downloaded from

https://opendata-ajuntament.barcelona.cat/data/en/dataset/arbrat-viari/resource/7613c24b-33ab-4671-8f01-c509c653c645

https://opendata-ajuntament.barcelona.cat/data/en/dataset/arbrat-parcs/resource/c569ee47-f734-4659-aebc-3a7b9f453807

https://opendata-ajuntament.barcelona.cat/data/en/dataset/arbrat-zona/resource/2ab05031-0efa-4f60-836f-a125c6be0061

Sets height=6.0m and radius=3.5m, as these values are not given in the dataset.

arguments

path: Path to the file with the dataset

extent: Extents.Extent, specifying a clipping range for the Dataset. Use X for lon and Y for lat.

returns

DataFrame conforming to the requirements needed to be considered a source of tree data.

TreeLoaders.load_valencia_trees — Method

load_valencia_trees(path; extent=nothing)

Loads tree data from valencia. Downloaded from https://valencia.opendatasoft.com/explore/dataset/arbratge-arbolado/export/ (using GeoJSON).

Sets height=6.0m and radius=3.5m, as these values are not given in the dataset.

arguments

path: Path to the file with the dataset

extent: Extents.Extent, specifying a clipping range for the Dataset. Use X for lon and Y for lat.

returns

DataFrame conforming to the requirements needed to be considered a source of tree data.