Loading New York Trees

Introduction

We load the tree data from the open data server of the city of new york, available here: new york tree data.

Since there are no height and radius data available, we set them constant to height=6m and radius=3.5m. These values are the mean values found in the Nottingham dataset.

API

TreeLoaders.load_new_york_trees

TreeLoaders.load_new_york_trees — Method

load_new_york_trees(path; extent=nothing, borough=nothing)

Loads tree data from New York. Downloaded from https://data.cityofnewyork.us/Environment/2015-Street-Tree-Census-Tree-Data/uvpi-gqnh.

Sets height=6.0m and radius=3.5m, as these values are not given in the dataset.

arguments

path: Path to the file with the dataset

extent: Extents.Extent, specifying a clipping range for the Dataset. Use X for lon and Y for lat.

borough: filter trees by borough. Select from: ["Manhattan", "Brooklyn", "Queens", "Staten Island", "Bronx"]

returns

DataFrame conforming to the requirements needed to be considered a source of tree data.