Shadow Casting

Introduction

We load the tree data from the open data server of the city concil of nottingham, available here: nottingham tree data.

As there are a lot of duplicates in this dataset, we need to clean that up, which happens while loading load_nottingham_trees. During this process, we also drop a lot of irrelevant columns and create a few new ones.

Due to the architecture of the shadow casting, we provide a user defined function which calculates the relevant data for calculating each trees shadow (x, y, height, radius) from every row in the loaded dataset. This allows for custom datasets to not be massively manipulated on loading, but rather enables the users to provide their own handlers working for their data.

API

TreeLoaders.load_nottingham_trees

TreeLoaders.tree_param_getter_nottingham

TreeLoaders.load_nottingham_trees — Method

load_nottingham_trees(path; bbox=nothing)

Loads tree data from nottingham. Downloaded from https://maps164.nottinghamcity.gov.uk/server/rest/services/OpenData/OpenData/MapServer/91.

merges the considerable amount of duplicate trees in the original dataset by distance (trees closer than 1e-4m are considered equal)

arguments

path: Path to the file with the dataset

bbox: named tuple with (minlon, minlat, maxlon, maxlat), specifying a clipping range for the TreeLoaders

returns

DataFrame with multiple columns, most relevant :id, :lon, :lat, :pointgeom and multiple other ones with a lot of information about the tree geometry and metadata keys "center_lon" and "center_lat" describing the center of the bounding box of all trees contained.

TreeLoaders.tree_param_getter_nottingham — Method

tree_param_getter_nottingham(row)

calculates the basic properies needed in the shadow casting algorithm based on a row of the nottingham dataset (loaded with load_nottingham_trees).

returns

tuple with:

x: first coordinate of the pointgeom

y: second coordinate of the pointgeom

height of the center of the tree crown

r: radius of the tree crown